@@ -1,12 +1,11 @@
 {
   "name": "silverstripe-webauthn",
-  "main": "./client/src/boot/index.js",
   "author": "SilverStripe Ltd",
   "engines": {
-    "node": ">=10.x"
+    "node": ">=18.x"
   },
   "scripts": {
-    "build": "yarn && yarn test && yarn lint && NODE_ENV=production webpack -p --bail --progress",
+    "build": "yarn && yarn lint && yarn test && rm -rf client/dist/* && NODE_ENV=production webpack --mode production --bail --progress",
     "dev": "NODE_ENV=development webpack --progress",
     "watch": "NODE_ENV=development webpack --watch --progress",
     "css": "WEBPACK_CHILD=css npm run build",
@@ -19,32 +18,39 @@
     "lint-sass": "sass-lint client/src"
   },
   "dependencies": {
-    "@silverstripe/react-injector": "^0.1.2",
-    "classnames": "^2.2.6",
-    "isomorphic-fetch": "^2.2.1",
-    "prop-types": "^15.7.2",
-    "react": "^16.8.3",
-    "react-dom": "^16.8.3"
+    "classnames": "^2.3.2",
+    "isomorphic-fetch": "^3.0.0",
+    "prop-types": "^15.8.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "@silverstripe/eslint-config": "^0.0.6",
-    "@silverstripe/webpack-config": "^1.1.0",
-    "babel-jest": "^23.6.0",
-    "copy-webpack-plugin": "^4",
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.10.0",
-    "jest-cli": "^23.6.0"
+    "@silverstripe/eslint-config": "^1.0.0-alpha6",
+    "@silverstripe/webpack-config": "^2.0.0-alpha5",
+    "babel-jest": "^29.2.2",
+    "copy-webpack-plugin": "^11.0.0",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.6",
+    "jest-cli": "^29.2.2",
+    "jest-environment-jsdom": "^29.3.1",
+    "react-16": "npm:react@^16.14.0",
+    "react-dom-16": "npm:react-dom@^16.14.0",
+    "webpack": "^5.74.0",
+    "webpack-cli": "^5.0.0"
   },
-  "babel": {
-    "presets": [
-      "env",
-      "react"
-    ],
-    "plugins": [
-      "transform-object-rest-spread"
-    ]
+  "resolutions": {
+    "colors": "1.4.0"
   },
+  "browserslist": [
+    "defaults"
+  ],
   "jest": {
+    "testEnvironment": "jsdom",
+    "moduleNameMapper": {
+      "^react-dom/client$": "react-dom-16",
+      "^react-dom((/.*)?)$": "react-dom-16$1",
+      "^react((/.*)?)$": "react-16$1"
+    },
     "roots": [
       "client/src"
     ],
@@ -65,7 +71,7 @@
       "**/tests/**/*-test.js?(x)"
     ],
     "transform": {
-      ".*": "babel-jest"
+      ".[jt]s[mx]?$": "babel-jest"
     }
   }
 }
