@@ -1,10 +1,36 @@
+/* eslint-disable max-len */
 import React from 'react';
 
 export default () => (
-  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 240 240"><defs>
-    {/* eslint-ignore-line max-len */}
-    <style>{'.a{fill:none;}.b{fill:#549ad3;}.c,.g{fill:#333a48;}.d{fill:#ffcf6e;}.e{clip-path:url(#a);}.f{fill:#d6e1ea;}.g{opacity:0.2;}.h{fill:url(#b);}.i{fill:#fff;}'}</style>
-    {/* eslint-ignore-line max-len */}
-    <clipPath id="a"><circle className="a" cx="120" cy="120" r="120" /></clipPath><linearGradient id="b" x1="-26.14" y1="68.86" x2="-11.09" y2="-38.22" gradientUnits="userSpaceOnUse"><stop offset="0.58" stopColor="#333a48" /><stop offset="0.59" stopColor="#424a5c" /></linearGradient></defs><title>U2F</title><circle className="b" cx="120" cy="120" r="120" /><path className="c" d="M127.91,93.4a1.22,1.22,0,0,0,1.21,1.22h35.61a3.77,3.77,0,0,0,3.56-4V74.33a3.77,3.77,0,0,0-3.56-4H129.12a1.22,1.22,0,0,0-1.21,1.22Zm34.38-13.32A2.22,2.22,0,1,1,160,82.3,2.24,2.24,0,0,1,162.29,80.08Z" /><path className="c" d="M114.18,76.44h15.35a0,0,0,0,1,0,0v9.12a3,3,0,0,1-3,3h-9.35a3,3,0,0,1-3-3V76.44A0,0,0,0,1,114.18,76.44Z" transform="translate(204.35 -39.35) rotate(90)" /><circle className="d" cx="146.94" cy="82.77" r="7.27" /><rect className="d" x="121.65" y="82.3" width="2.02" height="10.5" transform="translate(210.21 -35.11) rotate(90)" /><rect className="d" x="122.46" y="79.87" width="2.02" height="8.88" transform="translate(207.78 -39.15) rotate(90)" /><rect className="d" x="121.65" y="72.2" width="2.02" height="10.5" transform="translate(200.11 -45.21) rotate(90)" /><rect className="d" x="122.46" y="76.24" width="2.02" height="8.88" transform="translate(204.15 -42.78) rotate(90)" /><g className="e"><rect className="f" x="-157.14" y="54" width="285" height="192" rx="10" ry="10" /><rect className="c" x="-139.64" y="70" width="250" height="97" rx="2" ry="2" /><rect className="g" x="-62.64" y="175" width="96" height="62" rx="3" ry="3" /><rect className="c" x="-139.64" y="54" width="250" height="5" /><path className="h" d="M-149.51,51.5a2.52,2.52,0,0,1-2.25-1.41l-30.14-62a2.49,2.49,0,0,1,.13-2.42,2.47,2.47,0,0,1,2.12-1.17h330a2.47,2.47,0,0,1,2.12,1.17,2.47,2.47,0,0,1,.13,2.42l-30.14,62a2.5,2.5,0,0,1-2.25,1.41Z" /><path className="c" d="M150.37-13h0L120.23,49H-149.51l-30.14-62h330m0-5h-330a5,5,0,0,0-4.49,7.19l30.13,62a5,5,0,0,0,4.5,2.81H120.23a5,5,0,0,0,4.5-2.81l30.14-62a5,5,0,0,0-4.5-7.19Z" /></g><path className="i" d="M159,107.75a5.47,5.47,0,0,1,1.83.32,5.64,5.64,0,0,1,3.76,5.37c0,2.26,0,4.51,0,6.77v18.2a.6.6,0,0,0,.61.6.57.57,0,0,0,.5-.29c.89-1.49,2.34-3.08,4.49-3.08a6,6,0,0,1,4.32,1.5,5.45,5.45,0,0,1,1.43,2.18,1.28,1.28,0,0,0,1.2.86,1.2,1.2,0,0,0,.67-.21,5.81,5.81,0,0,1,3.2-1,5.33,5.33,0,0,1,2.77.76,5.81,5.81,0,0,1,2.43,3,5,5,0,0,1,.28,1.19,1.05,1.05,0,0,0,1,.93,1.07,1.07,0,0,0,.43-.09,5.53,5.53,0,0,1,2.4-.55h.08a5.57,5.57,0,0,1,5.45,4.66,17.12,17.12,0,0,1,.07,2.89v7.1c0,3.09-2.59,29-18,29H161.34l-1.1,0a4,4,0,0,1-1.42-.21A4.63,4.63,0,0,1,157,186a6,6,0,0,1-.5-.71L137,153.54a4.89,4.89,0,0,1,3.92-7.27h.22a5,5,0,0,1,3.86,1.86,8.76,8.76,0,0,1,.79,1.22l2.55,4.42,1.72,3a1.72,1.72,0,0,0,1.52.89,1.77,1.77,0,0,0,1.81-1.74c0-.73,0-1.26,0-1.42V113.52a5.62,5.62,0,0,1,5.61-5.77m0-5h0a10.6,10.6,0,0,0-10.61,10.75v30.94a10.12,10.12,0,0,0-7.28-3.18h-.47a9.9,9.9,0,0,0-8,14.77l0,.06,0,.05,19.47,31.75a12.19,12.19,0,0,0,4.79,4.45,8.84,8.84,0,0,0,3.29.57h17.65c18.94,0,23-27.28,23-34v-7.1c0-.4,0-.73,0-1a12.7,12.7,0,0,0-.16-2.75,10.55,10.55,0,0,0-10.37-8.8H190a10.89,10.89,0,0,0-3.7-3.75A10.45,10.45,0,0,0,181,134a10.69,10.69,0,0,0-2.34.25,10.59,10.59,0,0,0-.74-.76,10.91,10.91,0,0,0-7.74-2.85h-.6V118c0-1.51,0-3,0-4.55a10.61,10.61,0,0,0-7.09-10.06,10.38,10.38,0,0,0-3.5-.6Z" />
+  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 240 240">
+    <defs>
+      <style>{'.a{fill:none;}.b{fill:#549ad3;}.c,.g{fill:#333a48;}.d{fill:#ffcf6e;}.e{clip-path:url(#a);}.f{fill:#d6e1ea;}.g{opacity:0.2;}.h{fill:url(#b);}.i{fill:#fff;}'}</style>
+      <clipPath id="a">
+        <circle className="a" cx="120" cy="120" r="120" />
+      </clipPath>
+      <linearGradient id="b" x1="-26.14" y1="68.86" x2="-11.09" y2="-38.22" gradientUnits="userSpaceOnUse">
+        <stop offset="0.58" stopColor="#333a48" />
+        <stop offset="0.59" stopColor="#424a5c" />
+      </linearGradient>
+    </defs>
+    <title>U2F</title>
+    <circle className="b" cx="120" cy="120" r="120" />
+    <path className="c" d="M127.91,93.4a1.22,1.22,0,0,0,1.21,1.22h35.61a3.77,3.77,0,0,0,3.56-4V74.33a3.77,3.77,0,0,0-3.56-4H129.12a1.22,1.22,0,0,0-1.21,1.22Zm34.38-13.32A2.22,2.22,0,1,1,160,82.3,2.24,2.24,0,0,1,162.29,80.08Z" />
+    <path className="c" d="M114.18,76.44h15.35a0,0,0,0,1,0,0v9.12a3,3,0,0,1-3,3h-9.35a3,3,0,0,1-3-3V76.44A0,0,0,0,1,114.18,76.44Z" transform="translate(204.35 -39.35) rotate(90)" />
+    <circle className="d" cx="146.94" cy="82.77" r="7.27" />
+    <rect className="d" x="121.65" y="82.3" width="2.02" height="10.5" transform="translate(210.21 -35.11) rotate(90)" />
+    <rect className="d" x="122.46" y="79.87" width="2.02" height="8.88" transform="translate(207.78 -39.15) rotate(90)" />
+    <rect className="d" x="121.65" y="72.2" width="2.02" height="10.5" transform="translate(200.11 -45.21) rotate(90)" />
+    <rect className="d" x="122.46" y="76.24" width="2.02" height="8.88" transform="translate(204.15 -42.78) rotate(90)" />
+    <g className="e">
+      <rect className="f" x="-157.14" y="54" width="285" height="192" rx="10" ry="10" />
+      <rect className="c" x="-139.64" y="70" width="250" height="97" rx="2" ry="2" />
+      <rect className="g" x="-62.64" y="175" width="96" height="62" rx="3" ry="3" />
+      <rect className="c" x="-139.64" y="54" width="250" height="5" />
+      <path className="h" d="M-149.51,51.5a2.52,2.52,0,0,1-2.25-1.41l-30.14-62a2.49,2.49,0,0,1,.13-2.42,2.47,2.47,0,0,1,2.12-1.17h330a2.47,2.47,0,0,1,2.12,1.17,2.47,2.47,0,0,1,.13,2.42l-30.14,62a2.5,2.5,0,0,1-2.25,1.41Z" />
+      <path className="c" d="M150.37-13h0L120.23,49H-149.51l-30.14-62h330m0-5h-330a5,5,0,0,0-4.49,7.19l30.13,62a5,5,0,0,0,4.5,2.81H120.23a5,5,0,0,0,4.5-2.81l30.14-62a5,5,0,0,0-4.5-7.19Z" />
+    </g>
+    <path className="i" d="M159,107.75a5.47,5.47,0,0,1,1.83.32,5.64,5.64,0,0,1,3.76,5.37c0,2.26,0,4.51,0,6.77v18.2a.6.6,0,0,0,.61.6.57.57,0,0,0,.5-.29c.89-1.49,2.34-3.08,4.49-3.08a6,6,0,0,1,4.32,1.5,5.45,5.45,0,0,1,1.43,2.18,1.28,1.28,0,0,0,1.2.86,1.2,1.2,0,0,0,.67-.21,5.81,5.81,0,0,1,3.2-1,5.33,5.33,0,0,1,2.77.76,5.81,5.81,0,0,1,2.43,3,5,5,0,0,1,.28,1.19,1.05,1.05,0,0,0,1,.93,1.07,1.07,0,0,0,.43-.09,5.53,5.53,0,0,1,2.4-.55h.08a5.57,5.57,0,0,1,5.45,4.66,17.12,17.12,0,0,1,.07,2.89v7.1c0,3.09-2.59,29-18,29H161.34l-1.1,0a4,4,0,0,1-1.42-.21A4.63,4.63,0,0,1,157,186a6,6,0,0,1-.5-.71L137,153.54a4.89,4.89,0,0,1,3.92-7.27h.22a5,5,0,0,1,3.86,1.86,8.76,8.76,0,0,1,.79,1.22l2.55,4.42,1.72,3a1.72,1.72,0,0,0,1.52.89,1.77,1.77,0,0,0,1.81-1.74c0-.73,0-1.26,0-1.42V113.52a5.62,5.62,0,0,1,5.61-5.77m0-5h0a10.6,10.6,0,0,0-10.61,10.75v30.94a10.12,10.12,0,0,0-7.28-3.18h-.47a9.9,9.9,0,0,0-8,14.77l0,.06,0,.05,19.47,31.75a12.19,12.19,0,0,0,4.79,4.45,8.84,8.84,0,0,0,3.29.57h17.65c18.94,0,23-27.28,23-34v-7.1c0-.4,0-.73,0-1a12.7,12.7,0,0,0-.16-2.75,10.55,10.55,0,0,0-10.37-8.8H190a10.89,10.89,0,0,0-3.7-3.75A10.45,10.45,0,0,0,181,134a10.69,10.69,0,0,0-2.34.25,10.59,10.59,0,0,0-.74-.76,10.91,10.91,0,0,0-7.74-2.85h-.6V118c0-1.51,0-3,0-4.55a10.61,10.61,0,0,0-7.09-10.06,10.38,10.38,0,0,0-3.5-.6Z" />
   </svg>
 );
+/* eslint-enable max-len */
